split qwen36 moe residency shared types
