+++ crates/runner/src/main.rs
@@ -35,6 +35,7 @@ mod qwen36_moe_prefetch;
 mod qwen36_moe_prompt;
 mod qwen36_moe_residency;
 mod qwen36_moe_residency_pages;
+mod qwen36_moe_residency_types;
 mod qwen36_moe_session;
 mod qwen36_moe_spec_verify;
 mod qwen36_moe_speculative;
